+++ packages/mrm-task-travis/Changelog.md
@@ -1,3 +1,7 @@
+## Next
+
+* New option: maxNode.
+
 ## 1.0.0
 
 Initial release.